+++ blog-service/blog-server/src/main/java/com/wyz/blog/pojo/domain/BlogView.java
@@ -86,8 +86,8 @@ public class BlogView {
 	Integer collectionNum;
 
 	/**
-	 * 评分
-	 */
-	Integer score;
+     * 评分
+     */
+    Double score;
 
 }
